+++ assets/js/playList.js
@@ -1,25 +1,25 @@
 const playList = [
     {
         title: 'Aqua Caelestis',
-        src: 'assets/sounds/AquaCaelestis.mp3',
+        src: '../../assets/sounds/AquaCaelestis.mp3',
         duration: 39,
     },
     
     {
         title: 'Ennio Morricone',
-        src: 'assets/sounds/EnnioMorricone.mp3',
+        src: '../../assets/sounds/EnnioMorricone.mp3',
         duration: 97,
     },
     
     {      
         title: 'River Flows In You',
-        src: 'assets/sounds/RiverFlowsInYou.mp3',
+        src: '../../assets/sounds/RiverFlowsInYou.mp3',
         duration: 97
     },
 
     {      
         title: 'Summer Wind',
-        src: 'assets/sounds/SummerWind.mp3',
+        src: '../../assets/sounds/SummerWind.mp3',
         duration: 39
     },
