+++ index.html
@@ -49,6 +49,12 @@
       <div class="author"></div>
     </div>
   </footer>
-  <script type="module" src="index.js"></script>
+  <script src="assets/js/time.js"></script>
+  <script type="module" src="assets/js/greeting.js"></script>
+  <script type="module" src="assets/js/slider.js"></script>
+  <script src="assets/js/weather.js"></script>
+  <script type="module" src="assets/js/quotes.js"></script>
+  <script type="module" src="assets/js/player.js"></script>
+  <!-- <script type="module" src="index.js"></script> -->
 </body>
 </html>
